requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
greenlet==1.1.3
idna==3.3
numpy==1.23.2
pandas==1.4.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
